In C++ API, split internal implementation class into its own file(s), to follow more of a 1-class-per-file rule.
